qorpo/crypto/requirements.txt
aiohttp
ccxt
sqlalchemy
psycopg2
gunicorn
pytest-aiohttp